login user

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,16 @@ class TasksController extends Controller
      */
     public function index()
     {
-        // タスク一覧を取得
-        $tasks = Task::all();
-        
+        $tasks = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // タスクの一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+        }
+
         // タスク一覧ビューでそれを表示
-        return view('tasks.index',[
+        return view('tasks.index', [
             'tasks' => $tasks,
         ]);
     }
@@ -58,6 +63,7 @@ class TasksController extends Controller
         $task = new Task;
         $task->status = $request->status;  //追加
         $task->content = $request->content;
+        $task->user_id = \Auth::id();  //追加
         $task->save();
         
         // トップページへリダイレクトさせる
@@ -92,10 +98,16 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        // メッセージ編集ビューでそれを表示
-        return view('tasks.edit',[
-           'task' => $task, 
-        ]);
+        //タスクを保有しているユーザーかチェック
+        if ($task->user_id == \Auth::id()){
+            // メッセージ詳細ビューでそれを表示
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }else{
+            // トップページへリダイレクトさせる
+            return redirect('/');            
+        }
     }
 
     /**
@@ -115,10 +127,15 @@ class TasksController extends Controller
         
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        
+        
+        //タスクを保有しているユーザーかチェック
+        if ($task->user_id == \Auth::id()){
+            // タスクを更新
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -134,8 +151,11 @@ class TasksController extends Controller
     {
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        //タスクを保有しているユーザーかチェック
+        if ($task->user_id == \Auth::id()){
+            // メッセージを削除
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
